@@ -3,11 +3,11 @@
   "version": "13.0.1",
   "license": "MIT",
   "peerDependencies": {
-    "@angular/common": "^13.3.12",
-    "@angular/core": "^13.3.12"
+    "@angular/common": "^14.3.0",
+    "@angular/core": "^14.3.0"
   },
   "dependencies": {
-    "tslib": "^2.0.0"
+    "tslib": "^2.4.1"
   },
   "repository": {
     "type": "git",
